requirements.txt
flask
sentence-transformers
langchain
pinecone
PyMuPDF
python-dotenv
mistralai
